+++ lib/dsu/support/command_options/time_mneumonic.rb
@@ -22,16 +22,11 @@ module Dsu
         # a relative time mneumonic. Otherwise, it is optional.
         def time_from_mneumonic!(command_option:, relative_time: nil)
           validate_argument!(command_option: command_option, command_option_name: :command_option)
-          unless relative_time.nil?
-            validate_argument!(command_option: relative_time, command_option_name: :relative_time)
+          unless relative_time.nil? || relative_time.is_a?(::Time)
+            raise ArgumentError, "relative_time is not a Time object: \"#{relative_time}\""
           end
 
-          # if relative_time_mneumonic?(command_option) && !relative_time.nil?
-          #   # If command_option is a relative time mneumonic, we need to get the time
-          #   # relative to relative_time using ::Time.now, and use the command_option
-          #   # as the relative time.
-          #   return time_from_mneumonic!(command_option: relative_time, relative_time: command_option)
-          # end
+          relative_time ||= ::Time.now
 
           time_for_mneumonic(mneumonic: command_option, relative_time: relative_time)
         end
@@ -42,31 +37,19 @@ module Dsu
           mneumonic?(mneumonic) || relative_time_mneumonic?(mneumonic)
         end
 
+        # This method returns true if mneumonic is a valid relative
+        # time mneumonic.
+        def relative_time_mneumonic?(mneumonic)
+          return false unless mneumonic.is_a?(String)
+
+          mneumonic.match?(RELATIVE_REGEX)
+        end
+
         private
 
         # Returns a Time object from a mneumonic.
-        def time_for_mneumonic(mneumonic:, relative_time: nil)
-          # If relative_time is a relative time mneumonic, then we need to first
-          # convert mneumonic to a Time object first, so that we can calculate
-          # `relative_time.to_i.days.from_now(time)` to get the correct Time we
-          # need.
-          if relative_time_mneumonic?(relative_time)
-            time = time_for_mneumonic(mneumonic: mneumonic)
-            return relative_time_for(days_from_now: relative_time, time: time)
-          end
-
-          if mneumonic?(mneumonic) && mneumonic?(relative_time)
-            time = time_for_mneumonic(mneumonic: mneumonic)
-
-            # Simply return the time if relative_time is 'today'
-            # because 'today' relative to any time will always
-            # point to itself.
-            return time if today_mneumonic?(relative_time)
-
-            return time.public_send(relative_time)
-          end
-
-          time = ::Time.now
+        def time_for_mneumonic(mneumonic:, relative_time:)
+          time = relative_time
           if today_mneumonic?(mneumonic)
             time
           elsif tomorrow_mneumonic?(mneumonic)